+++ spec/lib/amz/competitive_pricing_spec.rb
@@ -14,24 +14,14 @@ RSpec.describe Amz::CompetitivePricing do
     end
   end
 
-  it "returns offers count product" do
+  it "returns total offer quantity" do
     VCR.use_cassette "amz/competitive_pricing" do
       asin = "B00BUFXI6M"
-      offers_count = 8
+      total_offers_count = 8
 
       result = described_class.new(api_keys_and_marketplace).for_asin(asin: asin)
 
-      expect(result.offers_count).to eq offers_count
+      expect(result.offers_count).to eq total_offers_count
     end
-  end
-
-  def api_keys_and_marketplace
-    attributes_for(:task)
-      .slice(:seller_id, :mws_auth_token)
-      .merge(marketplace: marketplace_us)
-  end
-
-  def marketplace_us
-    Peddler::Marketplace.find("US").id
   end
 end
