+++ spec/rails_helper.rb
@@ -19,6 +19,7 @@ end
 RSpec.configure do |config|
   config.include FactoryBot::Syntax::Methods
   config.include TaskHelper
+  config.include Amz
 
   config.use_transactional_fixtures = true
   config.infer_spec_type_from_file_location!
